+++ app/Config/Migration/1498311597_Migration_1.7.0.php
@@ -39,12 +39,28 @@ class Migration_1_7_0 extends PassboltMigration {
  */
 	public function after($direction) {
 		parent::after($direction);
+		$User = ClassRegistry::init('User');
+		$GroupUser = ClassRegistry::init('GroupUser');
+
 		// Move anonymous_statistics.php file from app/Config to /tmp
 		if (file_exists(APP . 'Config' . DS . 'anonymous_statistics.php')) {
 			if (!file_exists(AnonymousStatistic::ConfigFilePath)) {
 				mkdir(AnonymousStatistic::ConfigFilePath);
 			}
 			rename(APP . 'Config' . DS . 'anonymous_statistics.php', AnonymousStatistic::ConfigFilePath . DS . 'anonymous_statistics.php');
+		}
+
+		// Remove the deleted users from the groups they were member of.
+		$deletedUsers = $User->find('all', ['conditions' => [
+			'deleted' => 1
+		]]);
+		foreach ($deletedUsers as $deletedUser) {
+			$groupsUsers = $GroupUser->find('all', ['conditions' => [
+				'user_id' => $deletedUser['User']['id'],
+			]]);
+			foreach ($groupsUsers as $groupUser) {
+				$GroupUser->deleteGroupUser($groupUser);
+			}
 		}
 		return true;
 	}